+++ NEWS.md
@@ -5,6 +5,7 @@
 * Updated test coverage to use GHA
 * Fixed error handling for `pb_list()` for no release. 
 * `pb_list()` now respects the option `"piggyback.verbose"`
+* Fix download token handling [#88]
 
 
 # piggyback 0.1.4